# Conflicts:
#	src/TrueOrFalse.Frontend.Web/Views/Api/Search/SearchApiController.cs

+++ src/TrueOrFalse/Domain/Category/Ops/CategoryDeleter.cs
@@ -12,37 +12,48 @@ public class CategoryDeleter : IRegisterAsInstancePerLifetime
         _session = session;
     }
 
-    public void Run(Category category, bool forSetMigration = false)
+    public HasDeleted Run(Category category)
     {
         var categoryCacheItem = EntityCache.GetCategoryCacheItem(category.Id, getDataFromEntityCache: true);
+        var hasDeleted = new HasDeleted();
+
         if (categoryCacheItem.CachedData.ChildrenIds.Count != 0)
         {
             Logg.r().Error("Category can´t deleted it has children");
-            return;
+            hasDeleted.HasChildren = true;
+            return hasDeleted;
         }
 
-        if (!forSetMigration)
-            ThrowIfNot_IsLoggedInUserOrAdmin.Run(category.Creator.Id);
+        if (!Sl.SessionUser.IsInstallationAdmin || Sl.CurrentUserId != categoryCacheItem.Creator.Id)
+        {
+            hasDeleted.IsNotCreatorOrAdmin = true;
+            return hasDeleted;
+        }
 
-        Sl.UserActivityRepo.DeleteForCategory(category.Id);
+        _session.CreateSQLQuery("DELETE FROM relatedcategoriestorelatedcategories where Related_id = " + category.Id).ExecuteUpdate();
+        _session.CreateSQLQuery("DELETE FROM relatedcategoriestorelatedcategories where Category_id = " + category.Id).ExecuteUpdate();
+        _session.CreateSQLQuery("DELETE FROM categories_to_questions where Category_id = " + category.Id).ExecuteUpdate();
+        _session.CreateSQLQuery("DELETE FROM categories_to_sets where Category_id = " + category.Id).ExecuteUpdate();
 
-        if (forSetMigration)
-            Sl.CategoryRepo.DeleteWithoutFlush(category);
-        else
-            Sl.CategoryRepo.Delete(category);
+        Sl.UserActivityRepo.DeleteForCategory(category.Id); 
+        Sl.CategoryRepo.Delete(category);
 
         Sl.CategoryChangeRepo.AddDeleteEntry(category);
         Sl.CategoryValuationRepo.DeleteCategoryValuation(category.Id);
 
         ModifyRelationsEntityCache.DeleteIncludetContentOf(categoryCacheItem);
         CategoryRepository.UpdateCachedData(categoryCacheItem, CategoryRepository.CreateDeleteUpdate.Delete);
-        ModifyRelationsUserEntityCache.Delete(categoryCacheItem);
+        ModifyRelationsUserEntityCache.DeleteFromAllParents(categoryCacheItem);
         EntityCache.Remove(categoryCacheItem);
         UserCache.RemoveAllForCategory(category.Id);
+        hasDeleted.DeletedSuccessful = true;
+        return hasDeleted; 
+    }
 
-        _session.CreateSQLQuery("DELETE FROM relatedcategoriestorelatedcategories where Related_id = " + category.Id).ExecuteUpdate();
-        _session.CreateSQLQuery("DELETE FROM relatedcategoriestorelatedcategories where Category_id = " + category.Id).ExecuteUpdate();
-        _session.CreateSQLQuery("DELETE FROM categories_to_questions where Category_id = " + category.Id).ExecuteUpdate();
-        _session.CreateSQLQuery("DELETE FROM categories_to_sets where Category_id = " + category.Id).ExecuteUpdate();
+    public class HasDeleted
+    {
+        public bool HasChildren { get;  set; }
+        public bool IsNotCreatorOrAdmin { get; set; }
+        public bool DeletedSuccessful { get; set; }
     }
 }
